refactor common server functions
