+++ src/app/api/citizens/isCitizen.ts
@@ -1,20 +1,4 @@
-import "server-only";
+import { isCitizenForNamespace } from "../common/citizens/isCitizen";
 
-import { Prisma } from "@prisma/client";
-import prisma from "@/app/lib/prisma";
-
-export async function isCitizen(address: string) {
-  return prisma.$queryRaw<
-    {
-      address: string;
-    }[]
-  >(
-    Prisma.sql`
-    SELECT address
-    FROM center.address_metadata
-    WHERE kind = 'citizen' 
-    AND dao_slug = 'OP'
-    AND LOWER(address) = LOWER(${address});
-    `
-  );
-}
+export const isCitizen = (address: string) =>
+  isCitizenForNamespace(address, "optimism");
