+++ src/app/api/common/delegateStatement/delegateStatement.d.ts
@@ -1,0 +1,3 @@
+import { DelegateStatements } from "@prisma/client";
+
+export type DelegateStatement = Omit<DelegateStatements, 'createdAt' | 'updatedAt' | 'signature' | 'dao_slug'>;
