+++ src/app/api/common/delegations/delegation.d.ts
@@ -1,0 +1,12 @@
+import { OptimismDelegatees } from "@prisma/client";
+
+export type DelegateePayload = OptimismDelegatees;
+
+export type Delegation = {
+  from: string;
+  to: string;
+  allowance: string;
+  timestamp: Date | null;
+  type: "DIRECT" | "ADVANCED";
+  amount: "FULL" | "PARTIAL";
+};
