+++ src/app/api/common/metrics/metrics.d.ts
@@ -1,0 +1,5 @@
+export type Metrics = {
+  votableSupply: string;
+  totalSupply: string;
+  quorum: string;
+};
