+++ src/app/api/common/voting-power/votingPower.d.ts
@@ -1,0 +1,13 @@
+import {
+  OptimismAdvancedVotingPower,
+  OptimismVotingPower,
+} from "@prisma/client";
+
+export type AdvancedVotingPowerPayload = OptimismAdvancedVotingPower;
+export type VotingPowerPayload = OptimismVotingPower;
+
+export type VotingPowerData = {
+  directVP: string;
+  advancedVP: string;
+  totalVP: string;
+};
