+++ src/app/api/votableSupply/getVotableSupply.ts
@@ -1,12 +1,4 @@
-import prisma from "@/app/lib/prisma";
-import { DEPLOYMENT_NAME } from "@/lib/config";
+import { getVotableSupplyForNamespace } from "../common/votableSupply/getVotableSupply";
 
-export async function getVotableSupply() {
-  const votableSupply = await prisma[
-    `${DEPLOYMENT_NAME}VotableSupply`
-  ].findFirst({});
-  if (!votableSupply) {
-    throw new Error("No votable supply found");
-  }
-  return votableSupply.votable_supply;
-}
+export const getVotableSupply = () =>
+  getVotableSupplyForNamespace({ namespace: "optimism" });
